+++ Order.API/Order.API.UnitTest/WishListDetailFilterListTest.cs
@@ -78,7 +78,7 @@ namespace Order.API.UnitTest
         public OrderDTO GetOrderExists = new OrderDTO
         {
             Name = "WishList New Year",
-            Status = Shared.Entities.Enums.EnumOrderStatus.Active,
+            Status = (int)Shared.Entities.Enums.EnumOrderStatus.Active,
             Identifier = 55444
         };
